refactor: style 상수 정리
