+++ src/app/home/FollowSummary.tsx
@@ -2,7 +2,7 @@ import Banner from '@/components/Banner/Banner';
 import Icon from '@/components/Icon';
 import LevelProgressBar from '@/components/LevelStatus/LevelProgressBar';
 import Thumbnail from '@/components/Thumbnail/Thumbnail';
-import { gradientTextCss } from '@/constants/style';
+import { gradientTextCss } from '@/constants/style/text';
 import { css, cx } from '@styled-system/css';
 import { flex } from '@styled-system/patterns';
