+++ src/app/level/guide/page.tsx
@@ -8,7 +8,7 @@ import GrowthLevel from '@/app/level/guide/GrowthLevel';
 import LevelStatus from '@/components/LevelStatus/LevelStatus';
 import LoadingSpinner from '@/components/Loading/LoadingSpinner';
 import { LEVEL_SYSTEM } from '@/constants/level';
-import { defaultFadeInMotion } from '@/constants/style';
+import { defaultFadeInMotion } from '@/constants/style/animation';
 import { css, cx } from '@/styled-system/css';
 import { getLevel } from '@/utils/result';
 import { motion } from 'framer-motion';
