+++ src/constants/style/animation.ts
@@ -1,14 +1,6 @@
 import { css } from '@/styled-system/css';
 import { type Variants } from 'framer-motion';
 
-export const gradientTextCss = css({
-  animation: 'gradient 3s ease-in-out infinite',
-  backgroundSize: '150% 200%!',
-  backgroundClip: 'text!',
-  background: 'gradients.primary',
-  color: 'transparent',
-});
-
 export const fadeInAnimationCss = css({
   animation: `fadeIn  .7s ease-in`,
 });
